mysql_connector/requirements.txt
mysql-connector==2.2.9
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.4